fix bugs

+++ Source/Panel/CenterPanel/monitor_sub_panel.cpp
@@ -81,14 +81,13 @@ namespace zlpanel {
             auto deltaX = static_cast<float>(relativeTime.inSeconds() / timeInSeconds);
             auto area = image.getBounds().toFloat();
             deltaX = deltaX * area.getWidth();
-            deltaX = juce::roundToInt(deltaX);
+            deltaX = static_cast<float>(juce::roundToInt(deltaX));
         auto oldArea = area.withTrimmedLeft(deltaX - thickness).toNearestInt();
             lastInEndPoint = lastInEndPoint.translated(-deltaX - thickness * 0.1f, 0);
             lastOutEndPoint = lastOutEndPoint.translated(-deltaX - thickness * 0.1f, 0);
             lastDiffEndPoint = lastDiffEndPoint.translated(-deltaX - thickness * 0.1f, 0);
         auto oldImage = image.getClippedImage(oldArea);
         tempG.drawImageAt(image, static_cast<int>(-deltaX), 0);
-//            tempG.drawImageTransformed(image, juce::AffineTransform::translation(-deltaX, 0));
             // draw the new part
             if (!rmsIn.empty()) {
                 lock.enter();
@@ -122,7 +121,6 @@ namespace zlpanel {
             image.duplicateIfShared();
             // draw image to panel
             g.setOpacity(1.0f);
-//        g.drawImageAt(image, bound.toNearestInt().getX(), bound.toNearestInt().getY());
             g.drawImage(image, bound);
         }
     }
@@ -136,7 +134,6 @@ namespace zlpanel {
         lastInEndPoint = bound.getBottomRight();
         lastOutEndPoint = bound.getBottomRight();
         lastDiffEndPoint = bound.getTopRight();
-//        image = image.rescaled(getWidth() * upScaling, getHeight() * upScaling);
     }
 
     void MonitorSubPanel::setFontSize(float fSize) {
